+++ collections/in_mem/text_index.go
@@ -14,7 +14,8 @@ type InMemCollection struct {
 	mu             sync.RWMutex
 	collectionName string
 	lastInsertedID int64
-	TextMap        map[string]string                         // key: text
+	TextMap        map[string]string // key: text
+	LabelsMap      map[string][]string
 	IdMap          map[string]int64                          // key: postgres id
 	VectorIndexMap map[string]*interfaces.VectorIndexWrapper // searchMethod: vectorIndex
 }
@@ -23,6 +24,7 @@ func NewCollection(name string) *InMemCollection {
 	return &InMemCollection{
 		collectionName: name,
 		TextMap:        map[string]string{},
+		LabelsMap:      map[string][]string{},
 		IdMap:          map[string]int64{},
 		VectorIndexMap: map[string]*interfaces.VectorIndexWrapper{},
 	}
@@ -68,29 +70,37 @@ func (ti *InMemCollection) DeleteVectorIndex(ctx context.Context, searchMethod s
 	return nil
 }
 
-func (ti *InMemCollection) InsertTexts(ctx context.Context, keys []string, texts []string) error {
+func (ti *InMemCollection) InsertTexts(ctx context.Context, keys []string, texts []string, labelsArr [][]string) error {
 	if len(keys) != len(texts) {
 		return fmt.Errorf("keys and texts must have the same length")
 	}
-	ids, err := db.WriteCollectionTexts(ctx, ti.collectionName, keys, texts)
+
+	if len(labelsArr) != 0 && len(labelsArr) != len(keys) {
+		return fmt.Errorf("labels must have the same length as keys or be empty")
+	}
+
+	ids, err := db.WriteCollectionTexts(ctx, ti.collectionName, keys, texts, labelsArr)
 	if err != nil {
 		return err
 	}
 
-	return ti.InsertTextsToMemory(ctx, ids, keys, texts)
+	return ti.InsertTextsToMemory(ctx, ids, keys, texts, labelsArr)
 }
 
-func (ti *InMemCollection) InsertText(ctx context.Context, key string, text string) error {
-	id, err := db.WriteCollectionText(ctx, ti.collectionName, key, text)
+func (ti *InMemCollection) InsertText(ctx context.Context, key string, text string, labels []string) error {
+	id, err := db.WriteCollectionText(ctx, ti.collectionName, key, text, labels)
 	if err != nil {
 		return err
 	}
 
-	return ti.InsertTextToMemory(ctx, id, key, text)
+	return ti.InsertTextToMemory(ctx, id, key, text, labels)
 }
 
-func (ti *InMemCollection) InsertTextsToMemory(ctx context.Context, ids []int64, keys []string, texts []string) error {
+func (ti *InMemCollection) InsertTextsToMemory(ctx context.Context, ids []int64, keys []string, texts []string, labelsArr [][]string) error {
 
+	if len(labelsArr) != 0 && len(labelsArr) != len(keys) {
+		return fmt.Errorf("labels must have the same length as keys or be empty")
+	}
 	if len(ids) != len(keys) || len(ids) != len(texts) {
 		return fmt.Errorf("ids, keys and texts must have the same length")
 	}
@@ -99,16 +109,22 @@ func (ti *InMemCollection) InsertTextsToMemory(ctx context.Context, ids []int64,
 	defer ti.mu.Unlock()
 	for i, key := range keys {
 		ti.TextMap[key] = texts[i]
+		if len(labelsArr) != 0 {
+			ti.LabelsMap[key] = labelsArr[i]
+		}
 		ti.IdMap[key] = ids[i]
 		ti.lastInsertedID = ids[i]
 	}
 	return nil
 }
 
-func (ti *InMemCollection) InsertTextToMemory(ctx context.Context, id int64, key string, text string) error {
+func (ti *InMemCollection) InsertTextToMemory(ctx context.Context, id int64, key string, text string, labels []string) error {
 	ti.mu.Lock()
 	defer ti.mu.Unlock()
 	ti.TextMap[key] = text
+	if len(labels) != 0 {
+		ti.LabelsMap[key] = labels
+	}
 	ti.IdMap[key] = id
 	ti.lastInsertedID = id
 	return nil
@@ -135,6 +151,24 @@ func (ti *InMemCollection) GetTextMap(ctx context.Context) (map[string]string, e
 	ti.mu.RLock()
 	defer ti.mu.RUnlock()
 	return ti.TextMap, nil
+}
+
+func (ti *InMemCollection) GetLabels(ctx context.Context, key string) ([]string, error) {
+	ti.mu.RLock()
+	defer ti.mu.RUnlock()
+	return ti.LabelsMap[key], nil
+}
+
+func (ti *InMemCollection) GetLabelsMap(ctx context.Context) (map[string][]string, error) {
+	ti.mu.RLock()
+	defer ti.mu.RUnlock()
+	return ti.LabelsMap, nil
+}
+
+func (ti *InMemCollection) Len(ctx context.Context) (int, error) {
+	ti.mu.RLock()
+	defer ti.mu.RUnlock()
+	return len(ti.TextMap), nil
 }
 
 func (ti *InMemCollection) GetExternalId(ctx context.Context, key string) (int64, error) {