+++ collections/types.go
@@ -20,13 +20,34 @@ type collectionSearchResult struct {
 	Collection   string
 	SearchMethod string
 	Status       string
-	Objects      []collectionSearchResultObject
+	Objects      []*collectionSearchResultObject
 	Error        string
 }
 
 type collectionSearchResultObject struct {
 	Key      string
 	Text     string
+	Distance float64
+	Score    float64
+}
+
+type collectionClassificationResult struct {
+	Collection   string
+	SearchMethod string
+	Status       string
+	LabelsResult []*collectionClassificationLabelObject
+	Cluster      []*collectionClassificationResultObject
+	Error        string
+}
+
+type collectionClassificationLabelObject struct {
+	Label      string
+	Confidence float64
+}
+
+type collectionClassificationResultObject struct {
+	Key      string
+	Labels   []string
 	Distance float64
 	Score    float64
 }