+++ hostfunctions/collections.go
@@ -9,12 +9,17 @@ import (
 	wasm "github.com/tetratelabs/wazero/api"
 )
 
-func hostUpsertToCollection(ctx context.Context, mod wasm.Module, pCollectionName uint32, pKeys uint32, pTexts uint32) uint32 {
+func hostUpsertToCollection(ctx context.Context, mod wasm.Module, pCollectionName uint32, pKey uint32, pText uint32) uint32 {
+	return hostUpsertToCollectionV2(ctx, mod, pCollectionName, pKey, pText, 0)
+}
+
+func hostUpsertToCollectionV2(ctx context.Context, mod wasm.Module, pCollectionName uint32, pKeys uint32, pTexts uint32, pLabels uint32) uint32 {
 	var collectionName string
 	var keys []string
 	var texts []string
+	var labels [][]string
 
-	err := readParams3(ctx, mod, pCollectionName, pKeys, pTexts, &collectionName, &keys, &texts)
+	err := readParams4(ctx, mod, pCollectionName, pKeys, pTexts, pLabels, &collectionName, &keys, &texts, &labels)
 
 	if err != nil {
 		logger.Err(ctx, err).Msg("Error reading input parameters.")
@@ -23,13 +28,13 @@ func hostUpsertToCollection(ctx context.Context, mod wasm.Module, pCollectionNam
 
 	}
 
-	mutationRes, err := collections.UpsertToCollection(ctx, collectionName, keys, texts)
+	mutationRes, err := collections.UpsertToCollection(ctx, collectionName, keys, texts, labels)
 	if err != nil {
 		logger.Err(ctx, err).Msg("Error upserting to collection.")
 		return 0
 	}
 
-	offset, err := writeResult(ctx, mod, *mutationRes)
+	offset, err := writeResult(ctx, mod, mutationRes)
 	if err != nil {
 		logger.Err(ctx, err).Msg("Error writing result.")
 		return 0
@@ -55,7 +60,7 @@ func hostDeleteFromCollection(ctx context.Context, mod wasm.Module, pCollectionN
 		return 0
 	}
 
-	offset, err := writeResult(ctx, mod, *mutationRes)
+	offset, err := writeResult(ctx, mod, mutationRes)
 	if err != nil {
 		logger.Err(ctx, err).Msg("Error writing result.")
 		return 0
@@ -85,7 +90,33 @@ func hostSearchCollection(ctx context.Context, mod wasm.Module, pCollectionName 
 		return 0
 	}
 
-	offset, err := writeResult(ctx, mod, *searchRes)
+	offset, err := writeResult(ctx, mod, searchRes)
+	if err != nil {
+		logger.Err(ctx, err).Msg("Error writing result.")
+		return 0
+	}
+
+	return offset
+}
+
+func hostNnClassifyCollection(ctx context.Context, mod wasm.Module, pCollectionName uint32, pSearchMethod uint32, pText uint32) uint32 {
+	var collectionName string
+	var searchMethod string
+	var text string
+
+	err := readParams3(ctx, mod, pCollectionName, pSearchMethod, pText, &collectionName, &searchMethod, &text)
+	if err != nil {
+		logger.Err(ctx, err).Msg("Error reading input parameters.")
+		return 0
+	}
+
+	classification, err := collections.NnClassify(ctx, collectionName, searchMethod, text)
+	if err != nil {
+		logger.Err(ctx, err).Msg("Error classifying.")
+		return 0
+	}
+
+	offset, err := writeResult(ctx, mod, classification)
 	if err != nil {
 		logger.Err(ctx, err).Msg("Error writing result.")
 		return 0
@@ -138,7 +169,7 @@ func hostRecomputeSearchMethod(ctx context.Context, mod wasm.Module, pCollection
 		return 0
 	}
 
-	offset, err := writeResult(ctx, mod, *mutationRes)
+	offset, err := writeResult(ctx, mod, mutationRes)
 	if err != nil {
 		logger.Err(ctx, err).Msg("Error writing result.")
 		return 0